% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handypea_io.R
\name{calc_fluorojip_handypea}
\alias{calc_fluorojip_handypea}
\title{Calculate fluorojip from HandyPEA file}
\usage{
calc_fluorojip_handypea(file)
}
\arguments{
\item{file}{Path to the HandyPEA CSV file.}
}
\description{
Calculate fluorojip from HandyPEA file
}
