% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fluorojip}
\alias{example_fluorojip}
\title{Example OJIP Data (HandyPEA)}
\format{
A data frame with columns for sample identification,
treatment groups, and fluorescence induction values.
}
\usage{
data(example_fluorojip)
}
\description{
A dataset containing raw fluorescence values used to demonstrate
the calculation of JIP-test parameters.
}
\keyword{datasets}
