% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fluorojip.R
\name{plot_3d_fluorojip}
\alias{plot_3d_fluorojip}
\alias{plot_param_3d}
\alias{plot_param_surface3d}
\title{Plot 3D Scatterplot of OJIP Parameters}
\usage{
plot_3d_fluorojip(
  res,
  params = c("Fv_Fm", "PI_abs", "area"),
  group_col = "treatment",
  normalize = TRUE
)

plot_param_3d(
  res,
  params = c("Fv_Fm", "PI_abs", "area"),
  group_col = "treatment",
  normalize = TRUE
)

plot_param_surface3d(
  res,
  params = c("Fv_Fm", "PI_abs", "area"),
  group_col = "treatment",
  normalize = TRUE
)
}
\arguments{
\item{res}{Data frame with OJIP results.}

\item{params}{Character vector of exactly 3 parameters to plot.}

\item{group_col}{Name of the column with treatment groups.}

\item{normalize}{Logical indicating if data should be normalized.}
}
\description{
Plot 3D Scatterplot of OJIP Parameters
}
