% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handypea_io.R
\name{read_handypea_csv}
\alias{read_handypea_csv}
\title{Read HandyPEA CSV export}
\usage{
read_handypea_csv(file)
}
\arguments{
\item{file}{Path to the HandyPEA CSV file.}
}
\description{
Read HandyPEA CSV export
}
