% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fluorojip.R
\name{calc_fluorojip}
\alias{calc_fluorojip}
\title{Compute chlorophyll a OJIP parameters}
\usage{
calc_fluorojip(df)
}
\arguments{
\item{df}{A data frame with at least: t_fm, area, fo (or o), fm (or p), j, i, s.}
}
\value{
A data frame with calculated OJIP parameters.
}
\description{
Compute chlorophyll a fluorescence OJIP / JIP-test parameters
from raw fluorescence summary data with noise protection.
}
