% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fluorojip.R
\name{calc_fluorojip_file}
\alias{calc_fluorojip_file}
\title{Wrapper to read CSV and calculate OJIP parameters}
\usage{
calc_fluorojip_file(file, sep = ";", dec = ".", ...)
}
\arguments{
\item{file}{Path to the CSV file.}

\item{sep}{Separator (default ";").}

\item{dec}{Decimal point (default ".").}

\item{...}{Additional arguments passed to read.csv.}
}
\description{
Wrapper to read CSV and calculate OJIP parameters
}
