% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handypea_io.R
\name{handypea_to_ojip}
\alias{handypea_to_ojip}
\title{Convert HandyPEA to OJIP summary}
\usage{
handypea_to_ojip(x)
}
\arguments{
\item{x}{A list object returned by read_handypea_csv.}
}
\description{
Convert HandyPEA to OJIP summary
}
