% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fluorojip.R, R/normalize_fluorojip.R
\name{write_normalized_jiptable}
\alias{write_normalized_jiptable}
\title{Write Normalized JIP Table}
\usage{
write_normalized_jiptable(df, file, ...)

write_normalized_jiptable(df, file, ...)
}
\arguments{
\item{df}{The data frame to save.}

\item{file}{The path where the file will be saved.}

\item{...}{Additional arguments passed to write.csv.}
}
\description{
Write Normalized JIP Table

Write Normalized JIP Table
}
